Make space use an empty list for the default value of trusted_ip_ranges.
If trusted_ip_ranges isn't set, Heroku defaults to [0.0.0.0]. However, terraform
itself gets the value in the state set to [], resulting in a newly created space's state being
out of sync.

It is not possible (AFAICT at this point) to specify a default value for a List. Thefore,
we have no way to specify that the state should default to [0.0.0.0] that terraform can understand.

This change will break backwards compatibility - we might want to consider also making the
trusted_ip_range a required attribute, since a default of [] might be confusing to users.