Skip heroku_review_app_config.deploy_target.id region lookup for space
The current code assumes that the heroku_review_app_config.deploy_target is a region and tries to resolve it to a name, but in the case of a deploy target that is a Private Space, it should remain as the id, as space names would violate the schema and cause breaking changes. This skips the name look up if the type is space.